+++ src/pages/UserActivity/ApplicantCard.jsx
